Move evict_inline.h and evict_conn.c to src/evict/
